+++ src/Models/CodingStyles.ts
@@ -1,0 +1,15 @@
+export class CodingStyles {
+    convertNonPrivateFieldsToProperties = true;
+    useLanguageKeywordsInsteadOfFrameworkTypes = true;
+
+    get anyEnabled(): boolean {
+        return this.convertNonPrivateFieldsToProperties
+            || this.useLanguageKeywordsInsteadOfFrameworkTypes;
+    }
+
+    constructor(init: Partial<CodingStyles>) {
+        Object.assign(this, init);
+    }
+
+    // ! TODO: implement this in next release
+}